+++ src/app/services/spacex-api.service.ts
@@ -3,8 +3,7 @@ import {HttpClient, HttpErrorResponse, HttpParams} from '@angular/common/http';
 import {Observable, of, throwError} from "rxjs/index";
 import {Company} from "../models/company";
 import {catchError, map, tap} from "rxjs/internal/operators";
-import {LaunchFilter} from "../models/launchFilter";
-import { Launch } from '../models/launch';
+import {Launch, LaunchFilter} from '../models/launch';
 import {Capsule, CapsuleDetails, CapsuleFilter} from "../models/capsule";
 import {Rocket} from "../models/rocket";
 import {Launchpad} from "../models/launchpad";
